Restructure classes and interfaces into new files
