+++ DesignPatterns.Abstract/IHotDrink.cs
@@ -1,0 +1,6 @@
+namespace DesignPatterns.Abstract;
+
+public interface IHotDrink
+{
+    void Consume();
+}
