+++ DesignPatterns.Abstract/IHotDrinkFactory.cs
@@ -1,0 +1,6 @@
+namespace DesignPatterns.Abstract;
+
+public interface IHotDrinkFactory
+{
+    IHotDrink Prepare(int amount);
+}
